SQI updates

+++ amplitudes_widths_prominences.py
@@ -41,6 +41,6 @@ def get_amplitudes_widths_prominences(data,fs,visualise=0):
             plt.axis('off')
             plt.show()
 
-        return np.nanmedian(prominences), np.nanmedian(half_widths)/fs
+        return float(np.nanmedian(prominences)), float(np.nanmedian(half_widths)/fs)
     else:
         return np.NaN, np.NaN
